+++ hutool-crypto/src/main/java/cn/hutool/crypto/asymmetric/BaseAsymmetric.java
@@ -31,54 +31,6 @@ public class BaseAsymmetric<T extends BaseAsymmetric<T>> extends BouncyCastleSup
 
 	// ------------------------------------------------------------------ Constructor start
 	/**
-	 * 构造，创建新的私钥公钥对
-	 * 
-	 * @param algorithm 算法
-	 */
-	public BaseAsymmetric(String algorithm) {
-		this(algorithm, (byte[]) null, (byte[]) null);
-	}
-
-	/**
-	 * 构造 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
-	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
-	 * 
-	 * @param algorithm 非对称加密算法
-	 * @param privateKey 私钥
-	 * @param publicKey 公钥
-	 * @since 3.1.1
-	 */
-	public BaseAsymmetric(AsymmetricAlgorithm algorithm, PrivateKey privateKey, PublicKey publicKey) {
-		this(algorithm.getValue(), privateKey, publicKey);
-	}
-
-	/**
-	 * 构造 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
-	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
-	 * 
-	 * @param algorithm 非对称加密算法
-	 * @param privateKeyBase64 私钥Base64
-	 * @param publicKeyBase64 公钥Base64
-	 */
-	public BaseAsymmetric(String algorithm, String privateKeyBase64, String publicKeyBase64) {
-		this(algorithm, Base64.decode(privateKeyBase64), Base64.decode(publicKeyBase64));
-	}
-
-	/**
-	 * 构造
-	 * 
-	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
-	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
-	 * 
-	 * @param algorithm 算法
-	 * @param privateKey 私钥
-	 * @param publicKey 公钥
-	 */
-	public BaseAsymmetric(String algorithm, byte[] privateKey, byte[] publicKey) {
-		init(algorithm, privateKey, publicKey);
-	}
-
-	/**
 	 * 构造
 	 * 
 	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
@@ -97,23 +49,6 @@ public class BaseAsymmetric<T extends BaseAsymmetric<T>> extends BouncyCastleSup
 	/**
 	 * 初始化<br>
 	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
-	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密<br>
-	 * 
-	 * @param algorithm 算法
-	 * @param privateKey 私钥
-	 * @param publicKey 公钥
-	 * @return this
-	 */
-	public T init(String algorithm, byte[] privateKey, byte[] publicKey) {
-		final PrivateKey privateKeyObj = (null == privateKey) ? null : SecureUtil.generatePrivateKey(algorithm, privateKey);
-		final PublicKey publicKeyObj = (null == publicKey) ? null : SecureUtil.generatePublicKey(algorithm, publicKey);
-
-		return init(algorithm, privateKeyObj, publicKeyObj);
-	}
-
-	/**
-	 * 初始化<br>
-	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
 	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密（签名）或者解密（校验）
 	 * 
 	 * @param algorithm 算法
@@ -122,7 +57,7 @@ public class BaseAsymmetric<T extends BaseAsymmetric<T>> extends BouncyCastleSup
 	 * @return this
 	 */
 	@SuppressWarnings("unchecked")
-	public T init(String algorithm, PrivateKey privateKey, PublicKey publicKey) {
+	protected T init(String algorithm, PrivateKey privateKey, PublicKey publicKey) {
 		this.algorithm = algorithm;
 
 		if (null == privateKey && null == publicKey) {